+++ resources/js/hub-ui.js
@@ -22,3 +22,4 @@
 
 export { initToast } from './modules/toast.js';
 export { initConfirmation } from './modules/confirmation.js';
+export { registerCounter } from './modules/counter.js';